backend/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
jinja2==3.0.1
pillow==9.5.0
reportlab==3.6.1
flask-cors==3.0.10
numpy==1.24.3